remove underscore dependency and pika/pack for node/web
