+++ pkg/dist-types/index.d.ts
@@ -1,0 +1,4 @@
+export function makepuzzle(): any;
+export function posfor(x: any,y: any,axis: any): any;
+export function ratepuzzle(puzzle: any,samples: any): any;
+export function solvepuzzle(board: any): any;
